+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Go-to-defintion/Hover for RSpec DSL methods (eg. context or it)
+
+### Changed
 - Update testing steps in developer documentation
 
 ## [0.5.3] - 2025-09-02